+++ History.md
@@ -1,3 +1,6 @@
+### vNext ###
+- Added "Assertion message is invalid" warning
+
 ### 0.8.0 ###
 - Added "To-do items count" (see options page)
 - Added "Mismatch between type and file name" warning